pyproject.toml
[project]
name = "energy-aware-llm"
version = "0.1.0"
description = "Energy-aware TinyLlama inference benchmarking on consumer hardware"
authors = [
    { name = "Suprawee Pongpeeradech", email = "suprawee@email.sc.edu" }
]
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "plotly>=5.22",
    "polars>=0.20",
    "pynvml>=11.5",
    "pyyaml>=6.0",
    "rich>=13.7"
]

[project.optional-dependencies]
analysis = ["matplotlib>=3.8", "seaborn>=0.13"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
